# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: iROC GmbH website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: doctorLogin
      tags: [auth]
      summary: Doctor login with provided password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctors:
    get:
      operationId: listTrainedDoctors
      tags: [doctors]
      summary: Get list of trained doctors
      parameters:
        - name: instrument
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of trained doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainedDoctor"
  /geocode-postal:
    get:
      operationId: geocodePostal
      tags: [doctors]
      summary: Geocode a postal code for nearby-doctor searches
      parameters:
        - name: postal
          in: query
          required: true
          schema:
            type: string
        - name: country
          in: query
          required: true
          schema:
            type: string
          description: Country name or ISO alpha-2 code
      responses:
        "200":
          description: Postal-code coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoCoordinates"
        "400":
          description: Invalid postal code or country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Postal code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/doctors:
    post:
      operationId: createTrainedDoctor
      tags: [admin]
      summary: Admin - add a trained doctor to the public list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainedDoctorInput"
      responses:
        "201":
          description: Doctor added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainedDoctor"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/doctors/{id}:
    patch:
      operationId: updateTrainedDoctor
      tags: [admin]
      summary: Admin - update a trained doctor's full record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainedDoctorInput"
      responses:
        "200":
          description: Doctor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainedDoctor"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrainedDoctor
      tags: [admin]
      summary: Admin - remove a trained doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required:
        - instrument
        - password
      properties:
        instrument:
          type: string
          enum: [spirecut, ministem]
        password:
          type: string
    LoginResponse:
      type: object
      required:
        - success
        - instrument
      properties:
        success:
          type: boolean
        instrument:
          type: string
    SessionInfo:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
        instrument:
          type: ["string", "null"]
    DoctorCertification:
      type: object
      required:
        - instrument
        - certifiedDate
      properties:
        instrument:
          type: string
          description: Product identifier, e.g. 'spirecut', 'ministem', or any future product
        certifiedDate:
          type: string
    TrainedDoctor:
      type: object
      required:
        - id
        - firstName
        - lastName
        - city
        - country
        - certifications
      properties:
        id:
          type: integer
        title:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        specialty:
          type: ["string", "null"]
        institutionName:
          type: ["string", "null"]
        city:
          type: string
        postalCode:
          type: ["string", "null"]
        country:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/DoctorCertification"
    GeoCoordinates:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
        lon:
          type: number
    TrainedDoctorInput:
      type: object
      required:
        - firstName
        - lastName
        - city
        - country
        - certifications
      properties:
        title:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        specialty:
          type: ["string", "null"]
        institutionName:
          type: ["string", "null"]
        city:
          type: string
        postalCode:
          type: ["string", "null"]
        country:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
          minimum: -90
          maximum: 90
        lon:
          type: ["number", "null"]
          minimum: -180
          maximum: 180
        certifications:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DoctorCertification"
...